+++ src/convertRequest.js
@@ -1,15 +1,56 @@
 import { stringify } from 'query-string'
 import { GET_ONE, DELETE, GET_MANY, UPDATE, GET_MANY_REFERENCE, CREATE, GET_LIST } from './const'
+import { getIdFieldName } from './convertResponse'
 
 
 export const sort = (field, order) => order === "ASC" ? "-".concat(field) : field;
 
+/*
+  Translate from React-Admin to data provider (moleculer)
+ */
+export const uncustomizeItemFields = (json, resource, userOptions) => {
+  if (json === undefined || userOptions === undefined)
+    return json
 
-export const createOne = (params, apiUrl, resource) => ({
+  let customJson =  Object.assign({}, json); // clone
+  if (userOptions.renameFields !== undefined && userOptions.renameFields[resource] !== undefined) {
+    // loop params
+    for ( const [oldFieldName,newFieldName] of Object.entries( userOptions.renameFields[resource] ) ) {
+      if (customJson[newFieldName] !== undefined){
+        customJson[oldFieldName] = customJson[newFieldName]
+        delete customJson[newFieldName]
+      }
+    }
+  }
+
+  return customJson
+}
+
+export const unconvertIdField = (data, idField) => {
+  // already an _id or no id : stay as-is
+  if (data.id === undefined || data[idField] !== undefined)
+    return data
+
+  let resultData =  Object.assign({}, data); // clone
+
+  // rename id as _id
+  resultData[idField] = resultData.id
+  delete resultData.id
+  return resultData
+}
+
+export const applyOptions = (params, resource, userOptions) => {
+  const idField = getIdFieldName(resource, userOptions)
+  return uncustomizeItemFields( unconvertIdField(params, idField), resource, userOptions)
+};
+
+
+
+export const createOne = (params, apiUrl, resource, userOptions) => ({
   url: `${apiUrl}/${resource}`,
   options: {
     method: "POST",
-    body: JSON.stringify(params.data)
+    body: JSON.stringify(applyOptions(params.data, resource, userOptions))
   }
 });
 
@@ -82,11 +123,11 @@ export const getOne = (params, apiUrl, resource) => ({
 });
 
 
-export const updateOne = (params, apiUrl, resource) => ({
+export const updateOne = (params, apiUrl, resource, userOptions) => ({
   url: `${apiUrl}/${resource}/${params.id}`,
   options: {
     method: "PUT",
-    body: JSON.stringify(params.data)
+    body: JSON.stringify(applyOptions(params.data, resource, userOptions))
   }
 });
 
@@ -96,9 +137,10 @@ export const updateOne = (params, apiUrl, resource) => ({
  * @param {string} type One of the constants appearing at the top of this file, e.g. 'UPDATE'
  * @param {string} resource Name of the resource to fetch, e.g. 'posts'
  * @param {object} params The data request params, depending on the type
+ * @param {object} userOptions
  * @returns {object} { url, options } The HTTP request parameters
  */
-export const convertRequest = (apiUrl, type, resource, params) => {
+export const convertRequest = (apiUrl, type, resource, params, userOptions) => {
   let httpRequest = {
     url: "",
     options: {}
@@ -106,7 +148,7 @@ export const convertRequest = (apiUrl, type, resource, params) => {
 
   switch (type) {
     case CREATE:
-      httpRequest = createOne(params, apiUrl, resource)
+      httpRequest = createOne(params, apiUrl, resource, userOptions)
       break
     case GET_ONE:
       httpRequest = getOne(params, apiUrl, resource)
@@ -121,7 +163,7 @@ export const convertRequest = (apiUrl, type, resource, params) => {
       httpRequest = getManyReference(params, apiUrl, resource)
       break
     case UPDATE:
-      httpRequest = updateOne(params, apiUrl, resource)
+      httpRequest = updateOne(params, apiUrl, resource, userOptions)
       break
     case DELETE:
       httpRequest = deleteOne(params, apiUrl, resource)